# perf(hash_agg): bench q17 + always compact