--- a/src/stream/benches/hash_agg.rs
+++ b/src/stream/benches/hash_agg.rs
@@ -15,6 +15,7 @@
 use criterion::{black_box, criterion_group, criterion_main, BatchSize, Criterion};
 use futures::executor::block_on;
 use futures::StreamExt;
+use serde_json::error::Category::Data;
 use risingwave_common::catalog::{Field, Schema};
 use risingwave_common::types::DataType;
 use risingwave_expr::expr::*;
@@ -27,9 +28,11 @@ use risingwave_stream::executor::{BoxedExecutor, PkIndices};
 use tokio::runtime::Runtime;
 
 fn bench_hash_agg(c: &mut Criterion) {
-    let rt = Runtime::new().unwrap();
+    let mut group = c.benchmark_group("Q17");
+    group.sample_size(10);
 
-    c.bench_function("benchmark_hash_agg", |b| {
+    let rt = Runtime::new().unwrap();
+    group.bench_function("benchmark_hash_agg", |b| {
         b.to_async(&rt).iter_batched(
             || setup_bench_hash_agg(MemoryStateStore::new()),
             |e| execute_executor(e),
@@ -38,25 +41,70 @@ fn bench_hash_agg(c: &mut Criterion) {
     });
 }
 
+/// This aims to mirror `q17`'s aggregator.
+/// We can include more executor patterns as needed.
 fn setup_bench_hash_agg<S: StateStore>(store: S) -> BoxedExecutor {
     // ---- Define hash agg executor parameters ----
-    let data_types = vec![DataType::Int64; 3];
-    let schema = Schema {
-        fields: vec![Field::unnamed(DataType::Int64); 3],
-    };
+    let fields = vec![
+        Field::with_name(DataType::Int64, "auction"),
+        Field::with_name(DataType::Varchar, "to_char"),
 
-    let group_key_indices = vec![0];
+        Field::with_name(DataType::Int64, "count"),
+        Field::with_name(DataType::Int64, "count_filter_10_000"),
+        Field::with_name(DataType::Int64, "count_filter_10_000_to_100_000"),
+        Field::with_name(DataType::Int64, "count_filter_10_000_to_100_000"),
+
+        Field::with_name(DataType::Int64, "min"),
+        Field::with_name(DataType::Int64, "max"),
+        Field::with_name(DataType::Int64, "avg"),
+        Field::with_name(DataType::Int64, "sum"),
+    ];
+
+    let data_types = fields.iter().cloned().map(|field| {
+        field.data_type
+    }).collect();
+
+    let schema = Schema { fields };
+
+    let group_key_indices = vec![0, 1];
 
     let append_only = false;
 
     let agg_calls = vec![
-        AggCall {
-            kind: AggKind::Count, // as row count, index: 0
+         AggCall {
+            kind: AggKind::Count,
             args: AggArgs::None,
             return_type: DataType::Int64,
             column_orders: vec![],
             append_only,
             filter: None,
+            distinct: false,
+         },
+         AggCall {
+            kind: AggKind::Count,
+            args: AggArgs::None,
+            return_type: DataType::Int64,
+            column_orders: vec![],
+            append_only,
+            filter: Some(build_from_pretty("(less_than:boolean $0:int8 10000:int8)").into()),
+            distinct: false,
+         },
+         AggCall {
+            kind: AggKind::Count,
+            args: AggArgs::None,
+            return_type: DataType::Int64,
+            column_orders: vec![],
+            append_only,
+            filter: Some(build_from_pretty("(and:boolean (greater_than_or_equal:boolean $0:int8 10000:int8) (less_than:boolean $0:int8 100000:int8))").into()),
+            distinct: false,
+         },
+         AggCall {
+            kind: AggKind::Count,
+            args: AggArgs::None,
+            return_type: DataType::Int64,
+            column_orders: vec![],
+            append_only,
+            filter: Some(build_from_pretty("(greater_than_or_equal:boolean $0:int8 100000:int8)").into()),
             distinct: false,
         },
         AggCall {
@@ -77,10 +125,28 @@ fn setup_bench_hash_agg<S: StateStore>(store: S) -> BoxedExecutor {
             filter: None,
             distinct: false,
         },
+        AggCall {
+            kind: AggKind::Sum,
+            args: AggArgs::Unary(DataType::Int64, 2),
+            return_type: DataType::Int64,
+            column_orders: vec![],
+            append_only,
+            filter: None,
+            distinct: false,
+        },
+        AggCall {
+            kind: AggKind::Sum,
+            args: AggArgs::Unary(DataType::Int64, 2),
+            return_type: DataType::Int64,
+            column_orders: vec![],
+            append_only,
+            filter: None,
+            distinct: false,
+        },
     ];
 
     // ---- Generate Data ----
-    let num_of_chunks = 1000;
+    let num_of_chunks = 1024;
     let chunk_size = 1024;
     let chunks = gen_data(num_of_chunks, chunk_size, &data_types);