--- a/src/stream/src/executor/test_utils.rs
+++ b/src/stream/src/executor/test_utils.rs
@@ -560,13 +560,19 @@ pub fn gen_data(
         for _ in 0..chunk_size {
             ops.push(Op::Insert);
         }
+        // Generate columns of this chunk.
         for data_type in data_types {
-            let mut data_gen =
-                FieldGeneratorImpl::with_number_random(data_type.clone(), None, None, SEED)
-                    .unwrap();
+            // FIXME(kwannoel): This is specific to q17, this should be a configurable parameter.
+            // varchar (overwrite default, we want to simulate a fixed date string.
             let mut array_builder = data_type.create_array_builder(chunk_size);
             for j in 0..chunk_size {
-                array_builder.append_datum(&data_gen.generate_datum(((i + 1) * (j + 1)) as u64));
+                if *data_type == DataType::Varchar {
+                    array_builder.append_datum(&Some("2022-02-02".into()));
+                } else {
+                    let mut data_gen = FieldGeneratorImpl::with_number_random(data_type.clone(), None, None, SEED).unwrap();
+                    let datum = data_gen.generate_datum(((i + 1) * (j + 1)) as u64);
+                    array_builder.append_datum(datum);
+                }
             }
             columns.push(array_builder.finish().into());
         }